hypercore: add example for hypercore usage
this shows how a hypercore can be used for both offline and
online execution of tests. the testdata runs in a stream, which
is memory efficent.

for preparing the hypercore we take the example candles, load
them into a local hypercore and run the example. in practice in
could also be a live streaming connection to an online hypercore,
both works transparently :)

that means, in practice it does not matter if a backtest or even
trading strategy is running online or offline