add `manualOnStart` flag
the current honey framework is usually running `onStart` and
`onEnd` callbacks for the user. this was useful, but with streams
we need more fine grained control. additionally the requirements
of the ui have grown.

this adds a flag for backwards compat that allows the user to call
`onStart` manually. it is hidden behind a flag for backwards
compat. with 2.0 we will remove that flag.

it is useful in these scenarios: we have multiple streams, e.g.
stream a, b and c. we want to run the same strategy on all three
input types, but the current framework would call `onStart` for
all of them.