candles: don't try to emulate trades from candles
the current backtesting code takes a candle and tries to create
a few trades from it, to "reverse engineer" the trades during that
candle period. it tires to calculate the trades that lead to the
current candle, and then submits them into the strategy that is
run.

1. this comes as a surprise when you have chosen candles as an
input source (and opted out of trades), because the backtester
will try to sample trades and run the `onTrade` handlers which
influence the result. this is a major difference to the live
execution of a strategy in the HF (`bfx-hf-exec-strategy`)

2. the used sampling / emulation was never accurate. it tried to
use a static minimum amount for calculating the bonds for the upper
and lower prices, but with heavy price fluctuations and different
prices between trading pairs (i.e. ADA/BTC) this does not make
sense. also depending on the strategy volumes and other values
might be important, by sampling them we disturb the creation of
accurate results. an example was the trade amount which was set
evenly to the same amount between all emulated trades